refactor(webauthn): unify registration DTOs, require full profile for passkeys
Extract RegistrationDetailsDTO as a shared base for RegisterRequestDTO
and WebAuthnRegistrationStartDTO: username, firstname, lastname and
email are mandatory for both password- and passkey-based registration
now, only the password field itself differs. The dead optional
password field on the passkey DTO is gone, and the derived
getDisplayName() helper is @JsonIgnore'd so it no longer leaks into
the generated API schema. Regenerated the OpenAPI spec and API client
to match.
